JTroller.com Justin Trudeau Log Rolling Game v1.0
Fixed paused game issue with how deltaTime was calculated. In V1, when the game was paused, lastRenderTime stopped being updated then the un-pause event causes deltaTime to be off. This resulted in the attempt to draw all the missed frames, which of course leads to a quick game over. Fixed.